+++ src/components/surveys/inquire/submitted/card/CardQuestionItem.jsx
@@ -1,13 +1,16 @@
 import React, { useState, useEffect } from "react";
-import { QuestionWrapper2, QuestionOption2, QuestionText2 } from "./styled";
-import { userState } from "../../../authentication/userState";
-import Error from "../participate/Error";
-import Loading from "../participate/Loading";
+import { QuestionWrapper2, QuestionOption2, QuestionText2 } from "../styled";
+import { userState } from "../../../../../authentication/userState";
+import Error from "../../../participate/Error";
+import Loading from "../../../participate/Loading";
 import axios from "axios";
 import { useRecoilValue } from "recoil";
-import { surveyForSubmitted } from "../../../atoms";
+import { surveyForSubmitted } from "../../../../../atoms";
 import * as Sentry from "@sentry/react";
-import { getAccessToken, getRefreshToken } from "../../../authentication/auth";
+import {
+  getAccessToken,
+  getRefreshToken,
+} from "../../../../../authentication/auth";
 import styled from "styled-components";
 
 const QWrapper = styled.div`
