+++ src/components/surveys/inquire/submitted/card/CardSubmitted.jsx
@@ -1,14 +1,14 @@
 import React, { useState, useRef } from "react";
 import { useRecoilState } from "recoil";
-import { surveyForSubmitted } from "../../../atoms";
+import { surveyForSubmitted } from "../../../../../atoms";
 import Dialog from "@mui/material/Dialog";
 import DialogActions from "@mui/material/DialogActions";
 import DialogContent from "@mui/material/DialogContent";
 import DialogContentText from "@mui/material/DialogContentText";
 import DialogTitle from "@mui/material/DialogTitle";
 import Button from "@mui/material/Button";
-import MultipleChoiceQuestionItem from "./MultipleChoiceQuestionItem";
-import PreviewSection from "../create/card/PreviewSection";
+import MultipleChoiceQuestionItem from "../general/MultipleChoiceQuestionItem";
+import PreviewSection from "../../../create/card/PreviewSection";
 import styled from "styled-components";
 import CardQuestionItem from "./CardQuestionItem";
